requirements.txt
requests
python-dotenv
pytest
